net: xfrm: Use uintptr_t for offload_handle
As per the documentation it is ok to store a context pointer
in struct xfmr_dev_offload->offload_handle. Thus use uintptr_t
instead of unsigned long.

+++ include/net/xfrm.h
@@ -161,7 +161,7 @@ struct xfrm_dev_offload {
 	 * is held.
 	 */
 	struct net_device	*real_dev;
-	unsigned long		offload_handle;
+	uintptr_t		offload_handle;
 	u8			dir : 2;
 	u8			type : 2;
 	u8			flags : 2;
